+++ development/Tests/Providers/AmazonBedrockTests.cs
@@ -3,7 +3,7 @@ using Amazon;
 using Microsoft.Extensions.DependencyInjection;
 using Secrets;
 
-namespace AgentFrameworkToolkit.Tests.PaidModels;
+namespace AgentFrameworkToolkit.Tests.Providers;
 
 public sealed class AmazonBedrockTests : TestsBase
 {