+++ development/Tests/Providers/AzureOpenAITests.cs
@@ -7,7 +7,7 @@ using Secrets;
 
 #pragma warning disable OPENAI001
 
-namespace AgentFrameworkToolkit.Tests.PaidModels;
+namespace AgentFrameworkToolkit.Tests.Providers;
 
 public sealed class AzureOpenAITests : TestsBase
 {