+++ development/Tests/Providers/CohereTests.cs
@@ -2,7 +2,7 @@ using AgentFrameworkToolkit.Cohere;
 using Microsoft.Extensions.DependencyInjection;
 using Secrets;
 
-namespace AgentFrameworkToolkit.Tests.FreeModels;
+namespace AgentFrameworkToolkit.Tests.Providers;
 
 public sealed class CohereTests : TestsBase
 {