+++ development/Tests/Providers/OpenAITests.cs
@@ -6,7 +6,7 @@ using Secrets;
 
 #pragma warning disable OPENAI001
 
-namespace AgentFrameworkToolkit.Tests.PaidModels;
+namespace AgentFrameworkToolkit.Tests.Providers;
 
 public sealed class OpenAITests : TestsBase
 {